requirements.txt
streamlit
pandas
plotly
google-generativeai
python-dotenv
SpeechRecognition
gtts
streamlit-drawable-canvas
requests
pillow
numpy
scipy
matplotlib
streamlit_mic_recorder
bcrypt
uuid
redis
